Remove tree specific code from parent's undo() method.
Following from the previous commit that added an `UndoEntry.from_tab()` so
that the TabbedBrowser could have a generic way of creating an undo entry,
this commit adds an `UndoEntry.restore_into_tab()` that lets the parent
TabbedBrowser handle restoring a tab without the TreeTabbedBrowser class
overriding the `undo()` method.

Mostly this just moves code around so that TreeTabbedBrowser doesn't have to
store the undo entries before calling the super method, and we don't have to
change the parent method to return something just for a child class. The core
to that is being able to call a method per undo entry instead of going through
all of them at once. We could have added a new helper function in
TabbedBrowser to handle that and overridden it in TreeTabbedBrowser. But I
think it is nice to tuck the undo related code into the undo related object
anyhow.

The code doesn't look like it has moved because I changed the variable names:
newtab -> tab and entry -> self.