Make TreeUndoEntry a subclass of UndoEntry
This has two benefits:

1. removes tree lines of duplication
2. means that the entries in `AbstractTab.undo_entry` are always of the
   same type. It's a bit odd that the undo entry classes are private but
   that attribute they are stored in is public. I think the queue was
   made public just recently so perhaps an oversight.

The `kw_only` arg to dataclass is because the parent class sets
`created_at` with a default argument and dataclass enumerates the
attributes of the child class after the parent. If you tell it all
attributes need to be passed as keywords it doesn't care about that.