Make dataclass inheritance with defaults work with py<310
I've made TreeUndoEntry a subclass of UndoEntry. But UndoEntry sets a
default value for the `created_at` attribute, which means if any of the
extra attributes defined on the child class don't define a default an
error is thrown complaining about non-keyword arguments after a keyword
only one. Python 3.10 adds a kw_only argument you can pass to the
dataclass decorator but it looks like we currently support python 3.8+.
See https://www.trueblade.com/blogs/news/python-3-10-new-dataclass-features

So a more compatible way to do it is to set `init=False` on the
attribute with a default, if you don't need to pass it in the
constructor. Which we don't because it's only set by the caller in
tests. Hence changing the tests to set it a different way. A bit uglier
but should be fine.

Also fix my "retains_sort_order" test from #5964 that wasn't actually
being run...