Begin upgrading to eslint9
Initial migration was just:

    import yaml
    with open('.eslintrc.yaml') as f:
      config=yaml.load(f.read(), yaml.CLoader)
    with open("eslint.config.js", "w") as f:
      json.dump(config, f)

Then add a new ignores key, figuring out how to import the default "all"
config, and the globals.

Added some ignore lines for undefined globals in the config file since
they are probably legitimately not available in a browser.

The `ignores` config item doesn't seem to be working, not sure why.

Other than that theres not many warnings from the files that are
actually supposed to be linted, so that's good!

I have no-idea how much of the rest of the config is or isn't working.