Delay fake-key events by 10ms
Similar to a209c86c55a4 "Delay QEvent driven mouse clicks by 10ms" it seems
that sending fake-key QEvents immediately after a navigation causes the events
to not be processed.

This is causing e2e tests in keyinput.feature to fail in a flaky, but
frequent, manner. This can also be resolved in some other ways like putting
a wait in the e2e tests like so:

        When I open data/keyinput/log.html
        And I wait 0.01s
        And I run :fake-key <Escape>

But relying on maintainers to remember to do that seems error prone. If this
10ms delay turns out to be something to get rid of we could try keep this
delay to be used in less cases:

1. add some magic to e2e tests where it compares the previous and current line
   and if it sees an open and a click-element or fake-key line next to each
   other it delays for a bit
2. in the application code in tab.send_event() store the time of last
   navigation and queue events for sending till after that

The affected tests in this case where:

    tests/end2end/features/test_keyinput_bdd.py::test_fakekey_sending_key_to_the_website
    tests/end2end/features/test_keyinput_bdd.py::test_fakekey_sending_special_key_to_the_website
    tests/end2end/features/test_keyinput_bdd.py::test_fakekey_sending_keychain_to_the_website