Revert "delay fake-key" and add waits in e2e tests instead
Previously (a209c86c55a4) I've added a delay in browser code before sending
events to the page to account with a race condition where events weren't
processed after navigating. Then I had to add extra checks to tests that had
tight timing requirements.

That was for click-element, this commit reverts an attempt at the same
strategy for fake-key and instead adds wait statements in the tests that where
hitting the original race condition (sending events "too soon" after a
navigation).

The reason for the different approach here is that after adding the delay in
fake-key I had to add an extra "wait for log line" message to a few tests with
tight timing requirements to watch for a Tab key press. That worked great for
webengine but it made some tests start failing on webkit. We don't seem to get
that log message on webkit. I've got no-idea why and frankly think I've spent
way too much time on just a handful of tests already.

It's unfortunate we have to add manually delays in the e2e tests. It makes me
think if anyone else adds a test in the future with this combination of steps
(open page, run fake-key) they'll run into the same issue and it'll be hard to
spot. Oh well, we'll deal with that when it comes up.

The tests that where failing on webkit are the ones in caret.feature
touched in this commit.

Reverts included in this commit:

Revert "Delay fake-key events by 10ms"

This reverts commit 9f050c7460c42f317ceaa20b320e97d371a2c0a0.

Revert "Wait for evidence of Tab key press in tests before proceeding"

This reverts commit d47d247941c4b1fe3adac0dae8be301b36aec85b.