Fix more type hints
New mypy 3.11 update got smarter and raise some issues, they appear to
be correct in all cases.

There are several `type: ignore[unreachable]` comments in conditionals
on `sys.stderr` being None, which were introduce in a comment
specifically to handle a case where `sys.stderr` could be None. So
presumable the ignore comments were just to shut mypy up when it was
making mistakes.

In `debug.py` it was complaining that the class handling branch was
unreachable, because the type hint was overly restrictive. We do indeed
handle both classes and objects.

`log.py` got some extra Optional annotations around a variable that
isn't set if `sys.stderr` is None.