Make failed subframe styling error matching more flexible
In the Qt6.8.0-beta2 release for some reason the error message now looks like;

    Failed to style frame: Failed to read a named property '_qutebrowser' from 'Window': Blocked a frame with origin "http://localhost:35549" from accessing a cross-origin frame.

It seems to have an extra "Failed to read a named property '_qutebrowser' from
'Window'" before the "Blocked a frame ..." bit. Seems like maybe a nested
exception situation? Not sure what's going on there but the exception is still
being caught, which is the point of the test.

Hopefully we don't have more issues with subframes cropping up...