Use pip compile to find new deps instead of install and freeze
In #8269 we saw some packages leaking into the pip freeze output that we
don't want in the requirements files (if setuptools isn't supposed to be
in there, why should its dependencies).
I've also greatly missed the comments that pip-compile puts in
requirements.txt files explaining where indirect dependencies come from.
So I took the opportunity to switch our tooling for updating and parsing
new dependencies and their versions to use pip-compile instead of `pip -U
install && pip freeze`.

It turned out to not be a big change because the pip freeze output is
largely compatible with requirements files (we are writing directly to
one after all). So we just need to switch what commands we are running
and triage any compatibility issues.

I chose `uv pip compile` instead of `pip-compile` because I like what uv
is doing (fast, aiming for compatibility, consolidating a confusing
ecosystem into a single tool). But pip-compile/tools should do the same
job if we want to go that route.

The biggest differences are:
* outputs normalized names: this generally results in a larger diff than
  otherwise (I propose we go through an regenerate all the requirements
  files in one go, and maybe add that commit to a blame ignore file) and
  requires our comparison logic to deal with normalized package names
  everywhere
* setuptools and pip not included in tox requirement file - not sure
  what to do about that yet, should they be in the .text-raw file?

TODO:
* remove support for pip_args?
* change markers in raw files to lower case? Ideally don't require, if a human
  can write them in any case and a robot can normalize we should do that. If
  if there are patterns with `._` in them as part of names, how do we handle
  that?
* pull out similar bits of `build_requirements*` methods
  * maybe make it so you can pass `requirements=None` to `init_venv` to
    make it not install stuff, install uv, do the uv invocation, gate
    all that behind a `--method="freeze|compile"` arg?
* add pip and setuptools to tox requirements file?
* basename requirements file names so they don't have
* `script_path/../../` in them in the annotated version
* add tests for the markers (with inputs of differing cases) to make
  sure they all still work
* update changelog check used in CI to normalize names too