Combine the two "build_requirements" methods
There was a fair bit of duplicate code, so I've pulled out the "take a
list of requirements, give me a new one" out to separate methods. The
stuff around parsing the output can stay common thankfully!

Even if we drop one of the methods this level of abstraction is probably
fine to keep.