Remove slow test with poor isolation
It would have been convenient to have an end2end test to make sure that
the output of the two requirements file compilation methods had the same
results. But I think there is a bit too much stuff going on in
`recompile_requirements` for that atm.

Making the local repo in the temp path and install fake packages there
works fine, although setting up the virtualenv isn't quick. But the
script currently installs pip, setuptools and wheel which means we have
to either 1) hit pypi.org in the tests to get them, or 2) download them at
the start of the test suite and put them in the local repo.

Not sure it's worth the effort to go down this rabbit hole when we
already have a dozen real requirements files to verify the change with.

I'm leaving this in the commit history because it was fun to get the
local repo working!