Enable delightful pip compile lineage annotations
I find these comments useful to show why packages are included in the
final compiled requirements files.

Required a small change to `recompile_requirements` to ignore these new
comment line (it was saying there was a new requirement with an empty
name that needed a changelog entry).

The addition of the `os.path.realpath()` is to clean up the paths of the
requirements files in the annotations, eg:

     check-manifest==0.49
    -    # via -r scripts/dev/../../misc/requirements/requirements-check-manifest.txt-raw
    +    # via -r misc/requirements/requirements-check-manifest.txt-raw