Move recursive tab-close code to treetabbedbrowser
This commit moves any tree tab specific behavior out of `tab_close()`
into TreeTabbedBrowser. I would like calling code not to have to deal
with tree specific logic if it doesn't have to. In this case all we
should need to do is tell the tabbed browser to close the tab and let
that handle the details. (An even better API might be to ask the tab to
close itself, maybe in the future.)

This makes `TabbedBrowser.close_tab()` more powerful but now some logic
is a bit awkwardly mixed between commands.py and treetabbedbrowser. Eg
`tab_close_prompt()` is now being called in two places.

One shortfall of this commit is that `--force` is no longer supported to
close pinned tabs without prompting. It should be able to be supported
by ... plumbing through yet another keyword argument! I'm not sure why
the tab selection logic is in commands.py, I feel like all of that
should be in TabbedBrowser too, and then the passed through kwargs would
hopefully be a bit cleaner because they wouldn't have to go through so
many layers.