Simplify tab-give --recursive implementation
This gets a bit simpler to read if we do things in one pass. Because of the
order we go through the tree we will always process a node's parent before we
get to it. So if we save the new tabs (or just the new nodes) we can set the
correct parent of a new tab right away. Additionally since we are always
appending tabs (same order as we traverse them) then we can just append it to
the list of children of the parent we found.

In this implementation we traverse through the subtree of tabs-to-be given and
create a new tab in the target window for each one. We construct the right
tree structure in the distination window by adding the new tabs to the
children list of the appropriate parent.

For some reason the `setCurrentWidget()` at the end didn't make the test fail,
even though it is checking for a specific active tab. Hmmm.