Support fixed index in `TreeTabbedBrowser.tabopen()`
There are various settings that the user can change to dictate where a
new tab will be opened (eg, first, next, last, prev). The non-tree
tabbed browser supports passing in an index to `tabopen()` to bypass
those settings and use absolute positioning for new tabs. This is used
for example when undoing a tab close or loading a session. So far for
tree tab enabled windows callers have had to manually manipulate the
tree to get absolute positioning. This is an attempt to provide absolute
positing with the existing interface, at least as much as we can.

How it works is that we still use the `related` and `sibling` flags to
tell where in the tree hierarchy the tab should be placed (represented
by selecting the new parent node in the code) but then use the index
passed in to figure out where the tab should go in the list of siblings.
We find the tab bar index of the sibling tabs and try to place the index
passed in relative to them.

This is a very loose interpretation of "absolute" positioning. We are
relying on the position of nodes in the tree having the same ordering as
the tabs in tha tab bar, which
`TreeTabWidget.update_tree_tab_positions()` tries to ensure, and I've
added an assert as a bonus. Callers could also just pass in `0` or
`math.inf` to force the tabs to the first
or last positions without having to mess with the current config.

Seems to be working for the test suite for now. We can always remove it
again if gaps in the logic come up.