Add docker compose file
I find it difficult to reproduce issues on old python versions, or when
archlinux brings in a newer build of Qt than I have and I can't be
bothered rebuilding.

I've been using docker for a while to help test in these environments
locally, but the docker command is a bit long and unwieldy. Docker
compose is a method I've used before to provide canned docker
invocation. We could probably get away with writing our own script for
it, but using standard components is good. This should work with
podman-compose too.

This commit encodes the docker command I've been using so that I, and
other people, can use it with a less arcane command line. Hopefully it
helps other people. There are still a few rough edges like for any
runs of the "other" service you will pretty much have to drop into a
bash shell and install tox and some X11 stuff. But the images we publish
work in one-liners (although they would be a little faster with a
tox/pip cache directory mounted from the host so they don't setup the
venv every time).

See also: the docker related hints in the contributor guidlines.